+++ tests/test_acescc.py
@@ -55,7 +55,7 @@ class TestACESccSerialize(util.ColorAssertsPyTest):
         (
             'color(--acescc 1.5 0.3 0)',
             {'color': True, 'fit': 'lch-raytrace'},
-            'color(--acescc 1.468 0.3 -0.00071)'
+            'color(--acescc 1.468 0.3 -0.00064)'
         ),
         ('color(--acescc 1.5 0.2 0)', {'fit': False}, 'color(--acescc 1.5 0.2 0)')
     ]